major simplification for auth state and TopNav cleanup
