+++ src/components/TopNavigation/TopNavigation.tsx
@@ -1,203 +1,79 @@
+import React, { useState, useContext } from "react";
 import {
-  useState,
-  useEffect,
-  useContext,
-  Dispatch,
-  SetStateAction,
-  useRef,
-} from "react";
-
-import {
-  onAuthStateChanged,
   signOut,
   signInWithPopup,
   GoogleAuthProvider,
+  User as FirebaseUser,
 } from "firebase/auth";
-import "./TopNavigation.css";
 import { GoogleContext } from "../../App";
 import { Link, useLocation } from "react-router-dom";
-import { HelpModal } from "../HelpModal/HelpModal";
 import { useLoading } from "../Spinner/useLoading";
 import { useToast } from "../Toast/ToastProvider";
+import "./TopNavigation.css";
+import { HelpLinkItem } from "./HelpLinkItem";
+import { MenuPopover } from "./MenuPopover";
+import { GoBackNavLinkItem } from "./GoBackNavLinkItem";
 
-const provider = new GoogleAuthProvider(); // Google Auth Provider
+// TODO: Should this be moved to Context?
+const provider = new GoogleAuthProvider();
 
-type User = {
-  email: string;
-  displayName: string;
-  photoURL: string | null;
-};
-
-// Type guard to check if authUser is valid
-function isValidAuthUser(
-  // eslint-disable-next-line @typescript-eslint/no-explicit-any
-  authUser: any,
-): authUser is User {
-  return (
-    typeof authUser === "object" &&
-    authUser !== null &&
-    typeof authUser.email === "string" &&
-    typeof authUser.displayName === "string" &&
-    typeof authUser?.photoURL === "string"
-  );
-}
-
-function mapAuthUserToInternalUser(authUser: unknown): User {
-  if (!isValidAuthUser(authUser)) {
-    throw new Error("Unexpected user type");
-  }
-
-  return {
-    email: authUser.email,
-    displayName: authUser.displayName,
-    photoURL: authUser.photoURL ?? null,
-  };
-}
-
-const GoBackNavLinkItem = ({ showBack }: { showBack: boolean }) => {
-  if (!showBack) {
-    return null;
-  }
-
-  return (
-    <Link className="go-back-link top-nav-link fourth-color" to="/">
-      <img
-        src="/back-arrow.png"
-        style={{ width: 20, height: 20, paddingRight: 16 }}
-      />
-      Back
-    </Link>
-  );
-};
-
-const HelpLinkItem = ({
-  showHelpState,
-}: {
-  showHelpState: [boolean, Dispatch<SetStateAction<boolean>>];
-}) => {
-  const [showHelp, setShowHelp] = showHelpState;
-  return (
-    <div style={{ display: "inline-flex" }}>
-      <img
-        alt="Help Icon"
-        src="/question.png"
-        onClick={() => setShowHelp(true)}
-        style={{ width: 20, height: 20, cursor: "pointer" }}
-      />
-      <HelpModal showHelp={showHelp} setShowHelp={setShowHelp} />
-    </div>
-  );
-};
-
-const MenuPopover = ({
-  children,
-  pathname,
-}: {
-  children: React.ReactElement;
-  pathname: string;
-}) => {
-  const [openMenu, setOpenMenu] = useState(false);
-
-  const popoverRef = useRef<HTMLDivElement>(null);
-
-  const handleClickOutside = (event: MouseEvent) => {
-    if (
-      popoverRef.current &&
-      !popoverRef.current.contains(event.target as Node)
-    ) {
-      setOpenMenu(false);
-    }
-  };
-
-  useEffect(() => {
-    document.addEventListener("mousedown", handleClickOutside);
-
-    return () => {
-      document.removeEventListener("mousedown", handleClickOutside);
-    };
-  });
-
-  useEffect(() => {
-    if (openMenu) {
-      setOpenMenu(false);
-    }
-    // eslint-disable-next-line react-hooks/exhaustive-deps
-  }, [pathname]);
-
-  return (
-    <div className="menu-popover-trigger-container">
-      <img
-        alt="More Options Icon"
-        onClick={() => setOpenMenu(!openMenu)}
-        src="/menu.png"
-      />
-      <div
-        className={`menu-popover-children-container ${openMenu ? "open" : "closed"}`}
-        ref={popoverRef}
-      >
-        {children}
-      </div>
-    </div>
-  );
-};
-
-export const TopNavigation = () => {
+export const TopNavigation: React.FC = () => {
   const [showHelp, setShowHelp] = useState<boolean>(false);
-  const [user, setUser] = useState<User | null>(null);
 
   const { auth } = useContext(GoogleContext);
 
+  const location = useLocation();
+
   const { startLoading, stopLoading } = useLoading();
 
-  const location = useLocation();
+  const [user, setUser] = useState<FirebaseUser | null>(null);
+
   const { showToast } = useToast();
 
-  // Monitor Firebase authentication state
-  useEffect(() => {
-    const unsubscribe = onAuthStateChanged(auth, (currentUser) => {
-      if (currentUser === null) {
-        setUser(null);
-
-        showToast("Logged out.", "success");
-      } else {
-        const user = mapAuthUserToInternalUser(currentUser);
-        setUser(user);
-
-        showToast(`Logged in as ${user.email}.`, "success");
-      }
-    });
-
-    return () => unsubscribe();
-  }, []);
+  const [isSigningIn, setIsSigningIn] = useState(false);
 
   const handleLogout = async () => {
     try {
       startLoading();
 
       await signOut(auth);
+
+      setUser(null);
+
+      showToast("Logged out.", "success");
     } catch (error) {
       console.error("Error signing out:", error);
+      showToast("Uh Oh! There was an issue logging out.", "error");
     } finally {
       stopLoading();
     }
   };
 
   const handleLogin = async () => {
+    if (isSigningIn) {
+      return; // Prevent multiple sign-in attempts
+    }
+
+    setIsSigningIn(true);
+
     try {
       startLoading();
 
-      await signInWithPopup(auth, provider);
+      const { user } = await signInWithPopup(auth, provider);
+
+      setUser(user);
+
+      showToast(`Logged in as ${user.displayName}.`, "success");
     } catch (error) {
       console.error("Error signing in:", error);
+      showToast("Uh Oh! There was an issue logging in.", "error");
     } finally {
       stopLoading();
+
+      setIsSigningIn(false);
     }
   };
 
-  // when user is unauthenticated
-  // 1. if not on home, back button
-  // 2. help icon
-  // 4. sign in button
   if (!user) {
     return (
       <div className="top-nav-container">
@@ -214,11 +90,6 @@ export const TopNavigation = () => {
     );
   }
 
-  // when user authenticated
-  // 1. if not on home, back button
-  // 2. help icon
-  // 3. sign out btn
-  // 4. profile icon
   return (
     <div className="top-nav-container">
       <div className="top-nav-left-col">
@@ -228,9 +99,6 @@ export const TopNavigation = () => {
       <div className="top-nav-right-col">
         <MenuPopover pathname={location.pathname}>
           <div className="popover-menu-container">
-            {/* <div className="popover-menu-item">
-              <Link to="account">Account</Link>
-            </div> */}
             <div className="popover-menu-item" onClick={handleLogout}>
               <a href="javascript:void(0);">Sign out</a>
             </div>
@@ -239,7 +107,7 @@ export const TopNavigation = () => {
         {user.photoURL && (
           <div className="profile-photo-container">
             <Link to="/">
-              <img alt="Profile Photo." src={user.photoURL} />
+              <img alt="Profile" src={user.photoURL} />
             </Link>
           </div>
         )}
